revert some old changes
