+++ frontend/src/components/info.tsx
@@ -1,19 +1,15 @@
-import { A } from '@solidjs/router'
 import { type Component, Show, useContext } from 'solid-js'
 import { formatDistanceToNow } from 'date-fns'
-import { metadataContext } from '..'
-
 const Info: Component = () => {
-  const { meta, update } = useContext(metadataContext)
 
   return (
     <div class="m-2 h-screen flex items-center justify-center">
       <div class="min-width flex flex-col gap-2">
 
-        <A class="flex items-center self-start justify-self-start gap-1 rounded-md px-2 btn" href='/'>
+        {/* <A class="flex items-center self-start justify-self-start gap-1 rounded-md px-2 btn" href='/'>
           <div class="border border-blue-500 rounded" i-carbon-arrow-left></div>
           <p>Back</p>
-        </A>
+        </A> */}
         <div class="flex flex-col gap-2 border rounded-xl p-4 shadow">
           <div class="mb-2 flex items-center gap-2">
             <h1 class="flex-shrink text-2xl font-bold">
@@ -23,11 +19,11 @@ const Info: Component = () => {
 
           <div class="stretch-item">
             <p> Version: </p>
-            <p> {meta.version} </p>
+            {/* <p> {meta.version} </p> */}
           </div>
           <div class="stretch-item">
             <p>Last update: </p>
-            <Show when={meta.updating} fallback={
+            {/* <Show when={meta.updating} fallback={
               <button class="flex items-center gap-2 px-2 unimportant btn" onclick={update}>
                 <span>{formatDistanceToNow(meta.last_update)}</span>
                 <div class="border border-blue-500 rounded" i-material-symbols-sync></div>
@@ -37,7 +33,7 @@ const Info: Component = () => {
                 <span class="tracking-wide">Updating..</span>
                 <div class="loading-spinner border border-blue-500 rounded" i-material-symbols-sync></div>
               </div>
-            </Show>
+            </Show> */}
           </div>
         </div>
       </div>
